@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 from collections.abc import MutableMapping, Sequence
 from copy import copy
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Generic, TypeVar, Union
+from typing import TYPE_CHECKING, Generic, TypeVar
 
 import numpy as np
 import pandas as pd
@@ -33,10 +33,10 @@ if TYPE_CHECKING:
     from .raw import Raw
 
 
-OneDIdx = Union[Sequence[int], Sequence[bool], slice]
+OneDIdx = Sequence[int] | Sequence[bool] | slice
 TwoDIdx = tuple[OneDIdx, OneDIdx]
 # TODO: pd.DataFrame only allowed in AxisArrays?
-Value = Union[pd.DataFrame, spmatrix, np.ndarray]
+Value = pd.DataFrame | spmatrix | np.ndarray
 
 P = TypeVar("P", bound="AlignedMappingBase")
 """Parent mapping an AlignedView is based on."""
@@ -376,9 +376,14 @@ PairwiseArraysBase._view_class = PairwiseArraysView
 PairwiseArraysBase._actual_class = PairwiseArrays
 
 
-AlignedMapping = Union[
-    AxisArrays, AxisArraysView, Layers, LayersView, PairwiseArrays, PairwiseArraysView
-]
+AlignedMapping = (
+    AxisArrays
+    | AxisArraysView
+    | Layers
+    | LayersView
+    | PairwiseArrays
+    | PairwiseArraysView
+)
 T = TypeVar("T", bound=AlignedMapping)
 """Pair of types to be aligned."""
 
@@ -408,9 +413,7 @@ class AlignedMappingProperty(property, Generic[T]):
 
         def fake(): ...
 
-        fake.__annotations__ = {
-            "return": Union[self.cls._actual_class, self.cls._view_class]
-        }
+        fake.__annotations__ = {"return": self.cls._actual_class | self.cls._view_class}
         return fake
 
     def __get__(self, obj: None | AnnData, objtype: type | None = None) -> T:
