@@ -40,7 +40,7 @@ class Raw:
         # construct manually
         if adata.isbacked == (X is None):
             # Move from GPU to CPU since it's large and not always used
-            if isinstance(X, (CupyArray, CupySparseMatrix)):
+            if isinstance(X, CupyArray | CupySparseMatrix):
                 self._X = X.get()
             else:
                 self._X = X
@@ -51,7 +51,7 @@ class Raw:
             self.varm = varm
         elif X is None:  # construct from adata
             # Move from GPU to CPU since it's large and not always used
-            if isinstance(adata.X, (CupyArray, CupySparseMatrix)):
+            if isinstance(adata.X, CupyArray | CupySparseMatrix):
                 self._X = adata.X.get()
             else:
                 self._X = adata.X.copy()
@@ -124,9 +124,9 @@ class Raw:
         oidx, vidx = self._normalize_indices(index)
 
         # To preserve two dimensional shape
-        if isinstance(vidx, (int, np.integer)):
+        if isinstance(vidx, int | np.integer):
             vidx = slice(vidx, vidx + 1, 1)
-        if isinstance(oidx, (int, np.integer)):
+        if isinstance(oidx, int | np.integer):
             oidx = slice(oidx, oidx + 1, 1)
 
         if not self._adata.isbacked:
